+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - adding delete db cluster endpoint action
 - adding change_subnets action to asg
 - adding has_subnets probe to asg
+- adding start_instances and reboot_instances actions to ec2
+- adding instance_state probe to ec2
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.10.0...HEAD